+++ src/loggers/index.ts
@@ -1,3 +1,4 @@
 export * from './BrowserLogger';
 export * from './PlainLogger';
 export * from './ChalkLogger';
+export * from './WebhookLogger';