+++ CHANGELOG.md
@@ -21,6 +21,19 @@ const loggers = [
 ];
 ```
 
+- `WebhookLogger` - a `LogFunctionFactory` to send serialized form of message to a specific web address in a POST request
+
+```ts
+import { WebhookLogger } from '@fliegwerk/logsemts';
+
+const loggers = [
+	// sends a serialized JSON form of the message to the specified address
+	WebHookLogger({ address: 'http://localhost:8080' })
+];
+```
+
+- documentation of library concept to the `README.md`
+
 ### Changed
 
 ### Deprecated